chore(Table): warn when highlighted cells lack aria-label
